Implement fetch method in BeedsClient
Basic implementation using URLSession to fetch data from the API.
Created `BreedsClientError` with possible errors, and I'm throwing
them in each situation accordingly.

The app scheme must have a "DOG_API_KEY" environment variable
configured with a real TheDogAPI key, so that it can be passed
in the header of the request.

The response is decoded directly into an array of `Breed`.