Create breeds search feature
Created a new BreedsSearch reducer, with an additional action
compared to BreedsList, called search, which takes a string
representing the search term as input.

The new action calls the new BreedsClient method, search, and
then calls the receiveBreeds action with the retrieved breeds.

Finally, there's a new BreedSearchCellView view, to display
breed information on a cell format. Created a new screen-like
view called BreedsSearchView, with a searchable field and a list
of breeds.

I've also added tests for the reducer.